<feature> add: Removing RB From the ContactRepository If Space Remove…

+++ Assets/com.frame.fpphysics2d/PureRuntime/Phase/TriggerExitDispatch2DPhase.cs
@@ -35,6 +35,9 @@ namespace JackFrame.FPPhysics2D {
 
             // Remove From Collision
             var collisionRepo = context.CollisionContactRepo;
+            if (a.IsTrigger && b.IsTrigger) {
+                return;
+            }
             var collisionEventCenter = context.CollisionEventCenter;
             ulong key = DictionaryKeyUtil.ComputeRBKey(a, b);
             if (collisionRepo.Remove(key)) {
